We don't add the /code dir anymore
The dockerfile gives an environment to build(out) the code in.
We don't need to have the code inside the Docker machine as we're not going to
ship the docker. So it 'only' needs to be a machine to build/develop the code
upon.

For sites, this might be different. But then, we could use a two-stage
setup. A dockerfile like this prepares a machine to do the "build" stage in
and that can output an actual production 'binary'. (A 'real' docker machine
without development tools, a .tgz, a .deb, whatever)